mnv2_first/gateware: rename getset to registerfile
Rename getset to registerfile. Ensure unit tests run.
