+++ proj/mnv2_first/gateware/test_registerfile.py
@@ -16,7 +16,7 @@
 from nmigen import Signal
 from nmigen.sim import Settle
 from nmigen_cfu import InstructionTestBase, TestBase
-from .getset import Xetter, RegisterSetter, GetSetInstruction
+from .getset import Xetter, RegisterSetter, RegisterFileInstruction
 
 
 class RegisterSetterTest(TestBase):
@@ -71,12 +71,12 @@ class _AccumulateXetter(Xetter):
                 m.next = "waiting"
 
 
-class GetSetInstructionTest(InstructionTestBase):
+class RegisterFileInstructionTest(InstructionTestBase):
     def create_dut(self):
         self.m.submodules['rs1'] = rs1 = RegisterSetter()
         self.m.submodules['rs2'] = rs2 = RegisterSetter()
         self.m.submodules['acc'] = adder = _AccumulateXetter()
-        return GetSetInstruction({
+        return RegisterFileInstruction({
             1: rs1,
             2: rs2,
             9: adder
